Create ColorCubeFlyby.cpp
Add keyboard control (r) to rotate the whole image.
Make the cube to stop. Key (s) to stop and (c) to continue. With the image still add keyboard control to go up (u) and down (d) the whole image.
With keyboard, add zoom in/zoom out (keys +: Zoom in, -: Zoom out).
Add more cubes with different colors, illumination, and brightness.
Add two vertical planes, one on the left and one on the right, and make the cubes bounce.